+++ app/src/main/java/com/example/a2zbilling/counter/BillList/ShowBillingHistoryFragmentAdapterForPriceQntyValue.java
@@ -34,7 +34,7 @@ public class ShowBillingHistoryFragmentAdapterForPriceQntyValue extends Recycler
 
         holder.textView_for_saleDetail_price.setText(currentsalesDetail.getSalingPrice());
         holder.textView_for_saleDetail_qty.setText(Double.toString(currentsalesDetail.getQuntity()));
-        String values = Double.toString(Integer.parseInt(currentsalesDetail.getSalingPrice()) * currentsalesDetail.getQuntity());
+        String values = Double.toString(Double.parseDouble(currentsalesDetail.getSalingPrice()) * currentsalesDetail.getQuntity());
         holder.textView_for_saleDetail_value.setText(values);
